renaming server directory as per: https://github.com/ardanlabs/gotraining/tree/master/topics/go/design/packaging
moving current code to api/

routing requests to fe/api

updating README.md
