+++ api/cmd/app.yaml
@@ -1,0 +1,6 @@
+service: api
+runtime: go
+api_version: go1.8
+handlers:
+- url: /.*
+  script: _go_app
