--- lib/controllers/controllers.go
@@ -1,4 +1,0 @@
-package controllers
-
-// ResponseObject is a simple mapping obejct
-type ResponseObject map[string]interface{}
