--- server/app.yaml
@@ -1,10 +1,0 @@
-runtime: go
-api_version: go1.8
-handlers:
-- url: /stylesheets
-  static_dir: stylesheets
-- url: /(.*\.(gif|png|jpg))$
-  static_files: static/\1
-  upload: static/.*\.(gif|png|jpg)$
-- url: /.*
-  script: _go_app
